@@ -1,6 +1,6 @@
 {
   "name": "mdalamin",
-  "version": "0.1.0",
+  "version": "3.0.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
@@ -15,18 +15,39 @@
     "@tsparticles/slim": "^3.0.3",
     "aos": "^2.3.4",
     "autoprefixer": "10.4.14",
+    "axios": "^1.7.9",
+    "bcryptjs": "^2.4.3",
+    "cloudinary": "^2.5.1",
+    "cloudinary-react": "^1.8.1",
+    "dotenv": "^16.4.7",
     "eslint": "8.40.0",
     "eslint-config-next": "13.4.2",
     "framer-motion": "^10.12.16",
-    "next": "13.4.2",
+    "jsonwebtoken": "^9.0.2",
+    "mini-css-extract-plugin": "^2.9.2",
+    "mongoose": "^8.9.4",
+    "multiparty": "^4.2.3",
+    "next": "^15.1.3",
+    "next-auth": "^4.24.11",
+    "node-2fa": "^2.0.3",
+    "otpauth": "^9.3.6",
     "postcss": "8.4.23",
-    "react": "18.2.0",
+    "qrcode": "^1.5.4",
+    "react": "^18.2.0",
     "react-animated-cursor": "^2.7.0",
     "react-countup": "^6.4.2",
-    "react-dom": "18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hot-toast": "^2.5.1",
     "react-icons": "^4.9.0",
+    "react-markdown": "^9.0.3",
+    "react-markdown-editor-lite": "^1.3.4",
     "react-simple-typewriter": "^5.0.1",
+    "react-sortablejs": "^6.1.4",
+    "speakeasy": "^2.0.0",
     "swiper": "^9.3.2",
     "tailwindcss": "3.3.2"
+  },
+  "devDependencies": {
+    "daisyui": "^4.12.23"
   }
 }
