@@ -1,4 +1,7 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {}
+const nextConfig = {
+    // reactStrictMode: true,
+};
 
-module.exports = nextConfig
+export default nextConfig;
+
